Misc. Changes.
Moved NonceKeeper into its own file from DigestAuthentication.scala; this is done since NonceKeeper uses
scala.collection.mutable while DigestAuthentication uses
scala.collection.immutable. Separating this avoids confusion.

Made the nonces map a mutable private val in NonceKeeper. We use this
map now for lock sync.
